Moved env_to_send_in_RPC function to libp2p/peer/peerstore.py
