+++ libp2p/kad_dht/peer_routing.py
@@ -10,7 +10,6 @@ import logging
 import trio
 import varint
 
-from libp2p import env_to_send_in_RPC
 from libp2p.abc import (
     IHost,
     INetStream,
@@ -23,6 +22,7 @@ from libp2p.peer.id import (
 from libp2p.peer.peerinfo import (
     PeerInfo,
 )
+from libp2p.peer.peerstore import env_to_send_in_RPC
 
 from .common import (
     ALPHA,
